Shield cache from object updates through use of clones
