+++ central/namespace/datastore/internal/store/postgres/gen.go
